Updated sell/buy actuation ranges
- This should serve to better filter false signals

The actuation ranges chosen for each asset uniquely is based on each of their MACD indicator values. The additions try to decrease false sell signals that are close to the zero line that can happen in a ranging market and also add a buy signal for crosses that are near and above the zero line that could also be decent entry points. The filtering values were determined by historical trends of each of the assets MACD chart making sure to not accidentally filter out the large sell signals. 

The goal is to filter as much of the ranging market noise as possible. These values can still very likely be optimized more.